add function game()
Added a function that invokes the playRound function 5 times using a for loop. Each time a game is complete, the result of the playerWin variable is used to increment playerWins or computerWins which allows us to keep track of the total result of the five games played. The function then prints the total winner using these results.